DEX-293 pt 2 - failing tests
