+++ tests/modules/sightings/resources/test_sighting_featured_asset_guid.py
